+++ frontend/webpack/static/scripts/pages/allpages.js
@@ -2,11 +2,9 @@
 
 var Page = (function (PageClass) {
 	
-	var _PAGE = new PageClass();
-	
 	function Page() {
 		
-		this.config = _PAGE.getConfig();
+		this.config = PageClass.getConfig();
 		
 		start(this.config);
 		
@@ -16,7 +14,7 @@ var Page = (function (PageClass) {
 		console.log('all pages start', config);
 	}
 	
-	return new Page();
+	return Page;
 	
 });